Improve code organization @version 0.2
Add new files to group similar code in one place for easy access
and tracking. Rename a few key terms across the project to streamline
naming scheme as well as simplify it. Apply other formatting changes.

Code has been tested successfully for all of the target file formats.
However, it is NOT working for files with no integer columns (and it
was not working earlier either). Nevertheless, the reason is known
that it is because of the use of untested timestep variable(s).