Improve row handling and add separate output files
Handle rows requested by the user in a more streamlined manner.
Make 0-indexed rows the primary means for performing calculations,
while functionality to use timesteps for calculations may be added
next.

Add separate output.cpp and output.h files for performing
calculations and printing results to the screen. Options to file
the results may be added next and simply copy the printer function
as a filer function for the time being.

Tests of row handling operations and printing of results are
successful.