Add dataRowTotal and do miscellaneous changes
Add a variable dataRowTotal to the class CmdArgs::FileData and move
the variables dataColTotal and dataDlmType from CmdArgs::Args to
CmdArgs::FileData, while making all the necessary modifications for
this change. Replace all variables, named or temporary, referring to
total rows with dataRowTotal.

Moreover, rename the functions Output::printData and
Output::fileData to Output::dataPrinter and Output::dataFiler to avoid
confusion with the classes CmdArgs::PrintData and CmdArgs::FileData.