Update classifyColumns for new and blank lines
Update the conditions in function classifyColumns to match the
function populateVectors so that they both deal with new and blank
lines in the same way.

Also add findTimestepColCandidates function for the upcoming
modification that will add only timestep column (first integer column
from the left) to the integer category and all other columns will be
treated as double columns on which calculations can be performed.