Add new features to cycles and FFT and debug
Add new features to cycles option so that the maximum and various types
of means of the peaks, crests and troughs of input data can be
calculated. Allow the user to file the complete peaks, crests and
troughs data. Suppress the filing of FFT results by default but allow
the user to choose to file the FFT data, if desired.

Remove various bugs, especially the bug in counting cycles when data
with constant values equal to mean is provided.

Update README to reflect the changes.